+++ tests/empty_urls.py
@@ -1,0 +1,3 @@
+"""A URLconf with none of the endpoints of the library routed."""
+
+urlpatterns = []